Prohibit an admin from deleting its own group, tidy up some files and the db

+++ backend/api/views.py
@@ -140,9 +140,12 @@ class UserGroupDetail(APIView):
             return Response("Group name already exists.", status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk, format=None):
-        userGroup = self.get_object(pk)
-        userGroup.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if not request.data['group_id'] == request.data['target_group_id']:
+            userGroup = self.get_object(pk)
+            userGroup.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response("You cannot delete your own group.", status=status.HTTP_403_FORBIDDEN)
 
 class DemoList(APIView):
     permission_classes = [permissions.IsAuthenticated, partial(OnlyAdminPermission, ['POST'])]
@@ -264,11 +267,6 @@ class InstanceDetail(APIView):
     def delete(self, request, pk, format=None):
         instance = self.get_object(pk)
         group_id = instance.group_id.id
-        #print(group_id)
-        #print(request.data["group_id"])
-        #host = instance.host
-        #pid = instance.pid
-        #if request.data["host"] == str(host) and request.data["pid"] == str(pid) and request.data["user_id"] == str(user_id):
         if request.data["group_id"] == str(group_id):
             try:
                 instance.delete()
